Moved XML to Arrivals code to its own class Moved opensaml code to it's own class They return to us not just 200 - but 201 and 204 status code They timeout in their url is regarding when they time us out to look for the head of arrivals in their message queue.
